Move SinkhornJacobianApply test
